+++ lib/core/easy_search.dart
@@ -13,17 +13,15 @@ class EasySearch<T> extends StatefulWidget {
     SearchResultSettings searchResultSettings,
     FilterPageSettings filterPageSettings,
     bool multipleSelect,
+    bool startWithValue,
     this.controller,
     this.onSearch,
     this.onChange,
     this.customItemBuilder,
-  })  : this.searchResultSettings = searchResultSettings != null
-            ? searchResultSettings
-            : const SearchResultSettings(),
-        this.filterPageSettings = filterPageSettings != null
-            ? filterPageSettings
-            : const FilterPageSettings(),
+  })  : this.searchResultSettings = searchResultSettings != null ? searchResultSettings : const SearchResultSettings(),
+        this.filterPageSettings = filterPageSettings != null ? filterPageSettings : const FilterPageSettings(),
         this.multipleSelect = multipleSelect ?? false,
+        this.startWithValue = startWithValue ?? false,
         super(key: key);
 
   ///
@@ -52,6 +50,50 @@ class EasySearch<T> extends StatefulWidget {
   /// Default value: false
   ///
   final bool multipleSelect;
+
+  ///
+  /// Parameter name: [startWithValue]
+  ///
+  /// When this property is
+  /// ```dart
+  /// startWithValue == true,
+  /// ```
+  /// you allow the [controller] to start with list of items filled
+  ///
+  /// and can also make use of [OnSearch] to make
+  ///
+  /// requests to the web server
+  ///
+  /// Default value: false
+  ///
+  /// Example:
+  /// ```dart
+  /// EasySearch(
+  ///   onSearch: (text) {
+  ///     print('Filter Query: $text');
+  ///     return getData(name: text);
+  ///   },
+  ///   startWithValue: true,
+  ///   multipleSelect: true,
+  ///   searchResultSettings: SearchResultSettings(
+  ///     padding: EdgeInsets.only(left: 8.0, top: 8.0, right: 8.0),
+  ///     label: LabelSettings.searchLabel(value: 'People'),
+  ///   ),
+  ///   filterPageSettings: FilterPageSettings(
+  ///   searchOnShow: false, //Run search automatically when opening the search screen
+  ///   ),
+  ///   controller: SearchItem(
+  ///     items: [
+  ///       Item(ModelExample(name: 'Tiago', age: 37), true),
+  ///       Item(ModelExample(name: 'Mel', age: 3), true),
+  ///       Item(ModelExample(name: 'Monique', age: 30), false),
+  ///       Item(ModelExample(name: 'Timothy', age: 0), false),
+  ///     ],
+  ///   ),
+  /// ),
+  /// ```
+  ///
+  final bool startWithValue;
 
   ///
   /// Parameter name: [controller]
@@ -96,8 +138,7 @@ class EasySearch<T> extends StatefulWidget {
   final CustomItemBuilder<T> customItemBuilder;
 
   @override
-  _EasySearchState<T> createState() =>
-      _EasySearchState<T>(controller: controller);
+  _EasySearchState<T> createState() => _EasySearchState<T>(controller: controller);
 }
 
 class _EasySearchState<T> extends State<EasySearch<T>> {
@@ -116,8 +157,10 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
       throw ('STOP: hOoo no Dev,\nThe controller cannot be null if OnSearch is also null');
     }
 
-    if (widget.onSearch != null && widget.controller != null) {
-      //Ignore offline list when OnSearch and Controller is not null
+    if (widget.onSearch != null && widget.controller != null && widget.startWithValue == false) {
+      //Ignore offline list when:
+      //OnSearch and Controller is not null
+      //and StartWithValue == false
       _controller = SearchItem(items: []);
     }
   }
@@ -168,18 +211,10 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
               padding: const EdgeInsets.only(top: 10.0),
               child: Container(
                 decoration: BoxDecoration(
-                  border: Border.all(
-                      color: widget
-                          .searchResultSettings.styleSearchPage.borderColor,
-                      width: _borderWidth),
-                  color: widget
-                      .searchResultSettings.styleSearchPage.backGroundColor
-                      .withOpacity(widget.searchResultSettings.styleSearchPage
-                          .backgroundColorOpacity),
-                  borderRadius: BorderRadius.circular(widget
-                      .searchResultSettings
-                      .styleSearchPage
-                      .borderRadiusCircular),
+                  border: Border.all(color: widget.searchResultSettings.styleSearchPage.borderColor, width: _borderWidth),
+                  color: widget.searchResultSettings.styleSearchPage.backGroundColor
+                      .withOpacity(widget.searchResultSettings.styleSearchPage.backgroundColorOpacity),
+                  borderRadius: BorderRadius.circular(widget.searchResultSettings.styleSearchPage.borderRadiusCircular),
                 ),
                 constraints: BoxConstraints(
                   minHeight: 33.0,
@@ -219,25 +254,18 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
                       if (result != null) {
                         _controller.listItems.updateList();
                         var itemsTemp = (result as SearchItem);
-                        if (itemsTemp == null ||
-                            itemsTemp.listItems == null ||
-                            itemsTemp.listItems.getListItems.length == 0) {
+                        if (itemsTemp == null || itemsTemp.listItems == null || itemsTemp.listItems.getListItems.length == 0) {
                           print('No items were selected');
                           _controller.clear();
                         } else {
                           //Check items selected == true;
 
-                          var selectedList = itemsTemp
-                              .getSelectedItems.getListItems
-                              .where((element) => element.selected)
-                              .toList();
+                          var selectedList =
+                              itemsTemp.getSelectedItems.getListItems.where((element) => element.selected).toList();
 
-                          if (_controller != null &&
-                              _controller.getSelectedItems.getListItems.length >
-                                  0) {
+                          if (_controller != null && _controller.getSelectedItems.getListItems.length > 0) {
                             for (var itemSelected in selectedList) {
-                              print(
-                                  'Item: ${itemSelected.item}  -  Selected: ${itemSelected.selected}');
+                              print('Item: ${itemSelected.item}  -  Selected: ${itemSelected.selected}');
                             }
                           }
                         }
@@ -250,15 +278,11 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
                   child: Row(
                     children: [
                       SizedBox(
-                        width: widget.searchResultSettings.styleSearchPage
-                            .paddingLeftSearchIcon,
+                        width: widget.searchResultSettings.styleSearchPage.paddingLeftSearchIcon,
                       ),
-                      widget.searchResultSettings.prefix ??
-                          const Icon(Icons.search,
-                              color: Colors.grey, size: 22),
+                      widget.searchResultSettings.prefix ?? const Icon(Icons.search, color: Colors.grey, size: 22),
                       SizedBox(
-                        width: widget.searchResultSettings.styleSearchPage
-                            .paddingRightSearchIcon,
+                        width: widget.searchResultSettings.styleSearchPage.paddingRightSearchIcon,
                       ),
                       Expanded(
                         child: Padding(
@@ -268,19 +292,13 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
                             builder: (_, __, ___) {
                               return this._controller?.hasSelection == true
                                   ? Wrap(
-                                      spacing: widget
-                                          .searchResultSettings
-                                          .styleSearchPage
-                                          .spacingBetweenItemsValue,
+                                      spacing: widget.searchResultSettings.styleSearchPage.spacingBetweenItemsValue,
                                       children: buildTextItem(
                                         context: context,
                                         searchItem: _controller,
                                       ),
                                     )
-                                  : _label(
-                                          widget.searchResultSettings.labelHint,
-                                          "search...") ??
-                                      Container();
+                                  : _label(widget.searchResultSettings.labelHint, "search...") ?? Container();
                             },
                           ),
                         ),
@@ -291,8 +309,7 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
                 ),
               ),
             ),
-            if (widget.searchResultSettings.label?.value != null &&
-                widget.searchResultSettings.label?.value?.isNotEmpty == true)
+            if (widget.searchResultSettings.label?.value != null && widget.searchResultSettings.label?.value?.isNotEmpty == true)
               Positioned(
                 left: 20,
                 top: 9.5,
@@ -308,8 +325,7 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
                   ),
                 ),
               ),
-            if (widget.searchResultSettings.label?.value != null &&
-                widget.searchResultSettings.label?.value?.isNotEmpty == true)
+            if (widget.searchResultSettings.label?.value != null && widget.searchResultSettings.label?.value?.isNotEmpty == true)
               Positioned(
                 left: 25,
                 child: ClipRRect(
@@ -328,13 +344,13 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
 
   //Call OnChange Method
   void callOnChange() {
+    if (widget.onChange == null) {
+      return;
+    }
+
     List<T> onChangeValues;
-    if (_controller != null &&
-        _controller.getSelectedItems != null &&
-        _controller.getSelectedItems.getListItems.length > 0) {
-      var selectedList = _controller.getSelectedItems.getListItems
-          .where((element) => element.selected)
-          .toList();
+    if (_controller != null && _controller.getSelectedItems != null && _controller.getSelectedItems.getListItems.length > 0) {
+      var selectedList = _controller.getSelectedItems.getListItems.where((element) => element.selected).toList();
 
       onChangeValues = [];
       selectedList.forEach(
@@ -352,13 +368,9 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
     for (var element in searchItem.getSelectedItems.getListItems) {
       listWidget.add(
         InkWell(
-          highlightColor:
-              widget.searchResultSettings.buildItemResult.highlightColor,
-          splashColor:
-              widget.searchResultSettings.buildItemResult.splashColor ??
-                  Theme.of(context).primaryColor,
-          borderRadius: BorderRadius.circular(widget.searchResultSettings
-              .buildItemResult.circularBackgroundBorderRadius),
+          highlightColor: widget.searchResultSettings.buildItemResult.highlightColor,
+          splashColor: widget.searchResultSettings.buildItemResult.splashColor ?? Theme.of(context).primaryColor,
+          borderRadius: BorderRadius.circular(widget.searchResultSettings.buildItemResult.circularBackgroundBorderRadius),
           onTap: () {
             print('The item - ${element.item.toString()} has been removed');
             element.selected = false;
@@ -369,70 +381,48 @@ class _EasySearchState<T> extends State<EasySearch<T>> {
             callOnChange();
           },
           child: Padding(
-            padding:
-                widget.searchResultSettings.buildItemResult.backgroundPadding,
+            padding: widget.searchResultSettings.buildItemResult.backgroundPadding,
             child: Container(
               decoration: BoxDecoration(
-                color: widget
-                    .searchResultSettings.buildItemResult.backgroundColor
-                    .withOpacity(widget.searchResultSettings.buildItemResult
-                        .backgroundOpacity),
-                borderRadius: BorderRadius.circular(widget.searchResultSettings
-                    .buildItemResult.circularBorderRadiusItem),
+                color: widget.searchResultSettings.buildItemResult.backgroundColor
+                    .withOpacity(widget.searchResultSettings.buildItemResult.backgroundOpacity),
+                borderRadius: BorderRadius.circular(widget.searchResultSettings.buildItemResult.circularBorderRadiusItem),
               ),
               child: Padding(
-                padding:
-                    widget.searchResultSettings.buildItemResult.itemPadding,
+                padding: widget.searchResultSettings.buildItemResult.itemPadding,
                 child: Wrap(
-                  spacing: widget
-                      .searchResultSettings.buildItemResult.spacingBetweenItem,
+                  spacing: widget.searchResultSettings.buildItemResult.spacingBetweenItem,
                   children: [
                     Text(
                       element.item.toString(),
                       style: TextStyle(
-                        fontSize: widget.searchResultSettings.buildItemResult
-                            .itemValue.fontSize,
-                        fontWeight: widget.searchResultSettings.buildItemResult
-                            .itemValue.fontWeight,
-                        letterSpacing: widget.searchResultSettings
-                            .buildItemResult.itemValue.letterSpacing,
-                        color: widget.searchResultSettings.buildItemResult
-                            .itemValue.color
-                            .withOpacity(widget.searchResultSettings
-                                .buildItemResult.itemValue.colorOpacity),
+                        fontSize: widget.searchResultSettings.buildItemResult.itemValue.fontSize,
+                        fontWeight: widget.searchResultSettings.buildItemResult.itemValue.fontWeight,
+                        letterSpacing: widget.searchResultSettings.buildItemResult.itemValue.letterSpacing,
+                        color: widget.searchResultSettings.buildItemResult.itemValue.color
+                            .withOpacity(widget.searchResultSettings.buildItemResult.itemValue.colorOpacity),
                       ),
                     ),
                     Padding(
-                      padding: const EdgeInsets.only(
-                          left: 8.0, top: 0.0, right: 8.0, bottom: 0.0),
+                      padding: const EdgeInsets.only(left: 8.0, top: 0.0, right: 8.0, bottom: 0.0),
                       child: Container(
                         width: 20,
                         height: 20,
                         decoration: BoxDecoration(
-                          color: widget.searchResultSettings.buildItemResult
-                              .removeItemIconBackgroundColor
-                              .withOpacity(widget
-                                  .searchResultSettings
-                                  .buildItemResult
-                                  .removeItemIconBackgroundOpacity),
+                          color: widget.searchResultSettings.buildItemResult.removeItemIconBackgroundColor
+                              .withOpacity(widget.searchResultSettings.buildItemResult.removeItemIconBackgroundOpacity),
                           shape: BoxShape.circle,
                         ),
                         child: Icon(
-                          widget.searchResultSettings.buildItemResult
-                              .removeItemIcon,
-                          color: widget.searchResultSettings.buildItemResult
-                                      .removeItemIconBackgroundColor !=
-                                  widget.searchResultSettings.buildItemResult
-                                      .removeItemIconColor
-                              ? widget.searchResultSettings.buildItemResult
-                                  .removeItemIconColor
-                                  .withOpacity(widget.searchResultSettings
-                                      .buildItemResult.removeItemIconOpacity)
-                              : Theme.of(context).primaryColor !=
-                                      widget.searchResultSettings
-                                          .buildItemResult.backgroundColor
-                                  ? Theme.of(context).primaryColor.withOpacity(
-                                      widget.searchResultSettings.buildItemResult.removeItemIconOpacity)
+                          widget.searchResultSettings.buildItemResult.removeItemIcon,
+                          color: widget.searchResultSettings.buildItemResult.removeItemIconBackgroundColor !=
+                                  widget.searchResultSettings.buildItemResult.removeItemIconColor
+                              ? widget.searchResultSettings.buildItemResult.removeItemIconColor
+                                  .withOpacity(widget.searchResultSettings.buildItemResult.removeItemIconOpacity)
+                              : Theme.of(context).primaryColor != widget.searchResultSettings.buildItemResult.backgroundColor
+                                  ? Theme.of(context)
+                                      .primaryColor
+                                      .withOpacity(widget.searchResultSettings.buildItemResult.removeItemIconOpacity)
                                   : Colors.red,
                           size: 20,
                         ),